fix bug of destroying object and add a HP bar

+++ Assets/Script/GameMainControl.cs
@@ -132,11 +132,11 @@ public class GameMainControl : MonoBehaviour
                 }
                 AnnouncementWindow.gameObject.SetActive(true);
                 StartCoroutine(AnnouncementWindow.HideWindow(3));
-                if (team1wincount >= Mathf.Round(WinUnit / 2))
+                if (team1wincount >= Mathf.Round((float)WinUnit / 2))
                 {
                     EndGame(Team1.myTeam);
                 }
-                else if (team2wincount >= Mathf.Round(WinUnit / 2))
+                else if (team2wincount >= Mathf.Round((float)WinUnit / 2))
                 {
                     EndGame(Team2.myTeam);
                 }
